Break out Resource and Output into their own files
