+++ sdk/nodejs/asset/archive.ts
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { Input } from "../resource";
 import { Asset } from "./asset";
 
 /**
