+++ sdk/nodejs/stackReference.ts
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { all, CustomResource, CustomResourceOptions, Input, Output, output } from "./resource";
+import { CustomResource, CustomResourceOptions } from "./resource";
+import { all, Input, Output, output } from "./output";
 
 /**
  * Manages a reference to a Pulumi stack. The referenced stack's outputs are available via the
